pyproject.toml
[project]
name = "refinery"
dynamic = ["version"]
license = { text = "Apache 2.0" }
classifiers = [
    "Private :: Do Not Upload",
]
description = "A service for refining documents."
readme = "README.md"
authors = [
    { name = "Elijah Passmore", email = "elijah.passmore@wattmaven.com" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.31",
    "click>=8.1.8",
    "fastapi>=0.115.11",
    "jsonschema>=4.23.0",
    "mistralai>=1.6.0",
    "openai>=1.72.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.2.1",
    "structlog>=25.2.0",
    "uvicorn>=0.34.0",
]

[project.urls]
Homepage = "https://wattmaven.com"
Source = "https://github.com/wattmaven/refinery"

[project.scripts]
refinery = "refinery.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "commitizen>=4.1.1",
    "httpx>=0.28.1",
    "lefthook>=1.10.10",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.4",
    "trio>=0.29.0",
]

[tool.hatch.version]
source = "vcs"
raw-options = { local_scheme = "no-local-version" }

[tool.ruff.lint]
# See https://docs.astral.sh/ruff/rules
select = [
    "I",
    "N801",
    "N802",
    "N803",
    "N804",
    "N805",
    "N806",
    "N807",
    "N811",
    "N812",
    "N813",
    "N814",
    "N815",
    "N816",
    "N817",
    "N818",
    "N999",
    "FAST001",
    "FAST002",
    "FAST003",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
update_changelog_on_bump = true
